requirements.txt
streamlit
google-generativeai
python-docx
Pillow
python-dotenv
PyYAML
PyMuPDF
